Ryddet i eksempelfiler og gjort tiltak for å gjøre koden mer lesbar.
