+++ packages/designmanual/dummydata/mockResourceData.js
@@ -1,0 +1,144 @@
+import { constants } from '@ndla/ui';
+
+export const ResourceData = [
+  {
+    id: '61532e39-7623-4c14-b544-350cc9c2cba4',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: 'https://cdn.pixabay.com/photo/2022/06/12/22/35/village-7258991_1280.jpg',
+      alt: '',
+    },
+    contentType: constants.contentTypes.ASSESSMENT_RESOURCES,
+  },
+  {
+    id: '61532e39-7623-4c14-b544-350cc9c2cba4',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.ASSESSMENT_RESOURCES,
+  },
+  {
+    id: '94c16735-7485-4f95-9591-f61df6291d87',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.EXTERNAL_LEARNING_RESOURCES,
+  },
+  {
+    id: 'ee3b8780-3056-4696-a33e-6c4609530a40',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.LEARNING_PATH,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.MULTIDISCIPLINARY_TOPIC,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.SOURCE_MATERIAL,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.SUBJECT,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.SUBJECT_MATERIAL,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.TASKS_AND_ACTIVITIES,
+  },
+  {
+    id: '787b735e-bcd8-4b22-ae37-349a18013c43',
+    title: 'My Resource',
+    link: '',
+    topics: ['Oppgave', 'Norsk', 'Muntlig'],
+    tags: ['tag', 'tag', 'tag'],
+    description:
+      "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been theindustry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to mak",
+    resourceImage: {
+      src: '',
+      alt: '',
+    },
+    contentType: constants.contentTypes.TOPIC,
+  },
+];
